+++ Builtins.ts
@@ -1,12 +1,10 @@
 import {
-  aliasMatches,
   charListToJsString,
   createError,
   createMonadPassthrough,
   createType,
   createTypedList,
   createTypedTuple,
-  createTypeVar,
   createVar,
   FeverType,
   FeverVar,
@@ -15,7 +13,6 @@ import {
   Primitives,
   recursiveToString,
   typeAssignableFrom,
-  TypeWeights,
 } from "./types";
 import {
   callFunctionByReference,
@@ -25,8 +22,24 @@ import {
   interpret,
   recreateExpressionWithVariables,
 } from "./interpreter";
-import { feverStringFromJsString, inferListType } from "./literals";
-import { Context } from "./vars";
+import { inferListType } from "./literals";
+import { Context } from "./Context.ts";
+import {
+  namedFilter,
+  namedMap,
+  namedMonadicFilter,
+  namedMonadicMap,
+  namedReduce,
+} from "./lib/HigherOrder.js";
+import { isNumeric, newOfType, stringify } from "./lib/CommonUtils.js";
+import {
+  addFunctionCase,
+  registerNewFunction,
+  serializeFunction,
+  typesFromSignature,
+} from "./lib/FunctionUtils.js";
+import { feverStringFromJsString, typeToString } from "./lib/StringUtils.js";
+import { morphTypes } from "./lib/TypeUtils.js";
 
 export function newFunction(
   arity: number,
@@ -510,7 +523,7 @@ export const builtins = {
       1,
       [Primitives.ANY],
       ([v], ctx) => {
-        console.log(charListToJsString(dispatchFunction("stringify", [v])));
+        console.log(charListToJsString(stringify(v)));
         return v;
       },
       {
@@ -826,555 +839,3 @@ export const builtins = {
     newFunction(1, [Primitives.ANY], ([item]) => createMonadPassthrough(item)),
   ],
 };
-
-export const morphTypes = (value, toType, ctx) => {
-  let intermediateValue = value;
-  const typePath = ctx.pathBetween(value.type, toType.value);
-  for (let i = 0; i < typePath.length - 1; i++) {
-    const transformation = ctx.morphisms[typePath[i]][typePath[i + 1]];
-    intermediateValue = callFunctionByReference(
-      transformation,
-      [intermediateValue],
-      ctx,
-      "morph_transform",
-    );
-  }
-  return intermediateValue;
-};
-
-export const standardLib = [
-  "copies = {of, count:#} => (1..count -> (of))",
-  "contains? = {lst:[], item} => (lst \\> (false, (item == @ | $), true))",
-  "unique_add = {lst:[], (contains?(lst, item))} => (lst)",
-  "unique_add = {lst:[], item} => (lst + item)",
-  "unique = {lst:[]} => (lst \\> ([], (unique_add($,@))))",
-  "is_whole? = {n:#} => (floor(n) == n)",
-  "sum = {lst:[]} => (lst \\> (0, ($ + @)))",
-  "min = {a:#, (a<=b):#} => (a)",
-  "min = {_:#, b: #} => (b)",
-  "min = {(len(lst) > 0):[]} => (lst \\> ((get(lst,0)), (?((@ < $), @, $))))",
-  "max = {a:#, (a >= b):#} => (a)",
-  "max = {_:#, b:#} => (b)",
-  "max = {(len(lst) > 0):[]} => (lst \\> ((get(lst,0)), (?((@ > $), @, $))))",
-  "slice = {lst:[], from:#} => (lst ~> (# >= from))",
-  "in_range? = {target:#, (lower <= target):#, (target < higher):#} => true",
-  "in_range? = {_:#, _:#, _: #} => false",
-  "slice = {lst:[], from:#, to:#} => (lst ~> (in_range?(#, from, to)))",
-  "head = {(len(lst) > 0):[]} => (get(lst,0))",
-  "tail = {lst:[]} => (fast_slice(lst,1))",
-  "set = {(unique(entries)):[]}",
-  "halve = {lst:[]} => ([fast_slice(lst,0,floor(len(lst) / 2)), fast_slice(lst, floor(len(lst) / 2 ))])",
-  "merge = {_:fn, [], l2:[]} => (l2)",
-  "merge = {_:fn, l1:[], []} => (l1)",
-  "merge = {compare:fn, l1:[], ((compare(get(l1,0),get(l2,0))) < 0):[]} => ((get(l1,0)) + (merge(compare,tail(l1),l2)))",
-  "merge = {compare:fn, l1:[], l2:[]} => ((get(l2,0)) + (merge(compare, l1, tail(l2))))",
-  "sort = {(len(lst) <= 1):[], _:fn} => (lst)",
-  "sort_helper = {split_list:[], compare:fn} => (merge(compare,sort(get(split_list,0),compare), sort(get(split_list,1),compare)))",
-  "sort = {lst:[], compare:fn} => (sort_helper(halve(lst), compare))",
-  "compare = {n1:#,(n1 < n2):#} => -1",
-  "compare = {n1:#,(n1 > n2):#} => 1",
-  "compare = {_:#,_:#} => 0",
-  "sort = {nums:[#]} => (sort(nums, compare))",
-  "reverse = {lst:[]} => (lst -> (get(lst, len(lst) - # - 1)))",
-  "sum = {str:string} => (str \\> (0, ($ + ord(@))))",
-  "hash = {str:string, mod:#} => (sum(str) % mod)",
-  "abs = {(a < 0):#} => (a * -1)",
-  "abs = {a:#} => (a)",
-  "all = {vs:[], mapper:fn} => (vs \\> (true,(mapper(@) & $)))",
-  "any = {vs:[], mapper:fn} => (vs \\> (false, (mapper(@) | $)))",
-  "reverse = {vs:[]} => (vs -> (get(vs, len(vs) - # - 1)))",
-  "first = {vs:[]} => (get(vs,0))",
-  "last = {vs:[]} => (get(vs, len(vs) - 1))",
-  "time_diff = {result: []} => (last(result) - first(result))",
-];
-
-// Defining adds doesn't work, ie.   "+ = {s:set, item} => (new(set, entries(s) + item))",
-
-export const registerBuiltins = (ctx: Context, extraBuiltins?: any) => {
-  const builtinReference = !!extraBuiltins ? extraBuiltins : builtins;
-  for (const functionName of Object.keys(builtinReference)) {
-    const patterns = builtinReference[functionName];
-    for (const pattern of patterns) {
-      registerNewFunction(functionName, ctx, pattern);
-    }
-  }
-
-  // Only use text std lib if no extra builtins were supplied here
-  if (!extraBuiltins) {
-    for (const line of standardLib) {
-      interpret(line);
-    }
-  }
-};
-
-const typesFromSignature = (signature) => {
-  return signature.value.map((i) => i.value[1].value);
-};
-
-const conditionsFromSignature = (signature) => {
-  const conditions = [];
-
-  const sigPatterns = signature.value;
-  const size = sigPatterns.length;
-  for (let i = 0; i < size; i++) {
-    const pattern = sigPatterns[i];
-    const condition = pattern.value[0];
-    const conditionName = condition.value[0];
-    const conditionExpression = condition.value[1];
-    conditions.push((argument, ctx) => {
-      ctx.assignValue(conditionName.value, argument);
-      const result = evaluate(conditionExpression.value);
-      if (result.type.baseName === "ERROR") {
-        return false;
-      }
-      return result.value;
-    });
-  }
-  return conditions;
-};
-
-const namesFromSignature = (signature) => {
-  return signature.value.map((i) => i.value[0].value[0].value);
-};
-
-const specificitiesFromSignature = (signature) => {
-  return signature.value.map((i) => i.value[0].value[2].value);
-};
-
-const registerNewFunction = (name, variables, functionObject, rawCase?) => {
-  let newCase;
-
-  if (rawCase) {
-    newCase = rawCase;
-  } else {
-    newCase = generateCaseFromNative(functionObject);
-  }
-  const named = variables.getOrNull(name);
-  if (!named) {
-    const newFunc = createVar([newCase], Meta.FUNCTION);
-    newFunc["invocations"] = [functionObject];
-    variables.assignValue(name, newFunc);
-    return newFunc;
-  }
-
-  named.invocations.push(functionObject);
-  named.value.push(newCase);
-  return named;
-};
-
-const nativeFunctionMessage = "<natively defined function>";
-
-const generateCaseFromNative = (functionObject) => {
-  const types = functionObject["types"];
-  const conditions = functionObject["conditions"];
-  const specificities = functionObject["specificities"];
-  const names = argNamesFromFunction(functionObject["function"].toString());
-
-  const patterns = [];
-  for (let i = 0; i < types.length; i++) {
-    const type = types[i];
-    const condition = conditions[i];
-    const pattern = createVar(
-      [
-        createVar(
-          [
-            createVar(names[i], Meta.STRING),
-            createVar(
-              condition.toString() === "() => true"
-                ? "true"
-                : nativeFunctionMessage,
-              Primitives.EXPRESSION,
-            ),
-            createVar(specificities[i], Primitives.NUMBER),
-          ],
-          Meta.CONDITION,
-        ),
-        createVar(type, Primitives.TYPE),
-      ],
-      Meta.PATTERN,
-    );
-    patterns.push(pattern);
-  }
-
-  return createVar(
-    [
-      createVar(patterns, Meta.SIGNATURE),
-      createVar(nativeFunctionMessage, Primitives.EXPRESSION),
-    ],
-    Meta.CASE,
-  );
-};
-
-const newOfType = (t, args, ctx: Context) => {
-  const typeVar = createTypeVar(t);
-  return dispatchFunction("new", [typeVar, ...args]);
-};
-
-const argNamesFromFunction = (functionBody) => {
-  const args = [];
-  let stack = "";
-  let inDestructure = false;
-  for (const letter of functionBody) {
-    if (inDestructure) {
-      switch (letter) {
-        case "]":
-          args.push(stack);
-          return args;
-        case ",":
-          args.push(stack);
-          stack = "";
-          break;
-        default:
-          stack += letter;
-          break;
-      }
-    } else if (letter === "[") {
-      inDestructure = true;
-    }
-  }
-  return args;
-};
-
-const serializeCase = (c) => {
-  const [signature, expression] = c.value;
-  const patterns = signature.value;
-
-  let caseText = patterns.length >= 1 ? "(" : "() => ";
-  for (let i = 0; i < patterns.length; i++) {
-    const pattern = patterns[i];
-    const [condition, type] = pattern.value;
-    const [varName, conditionExpr] = condition.value;
-    caseText +=
-      conditionExpr.value.value === true ? varName.value : "<condition>";
-    if (type.value.baseName !== "ANY") {
-      caseText +=
-        ":" +
-        (isAlias(type.value)
-          ? type.value.alias
-          : type.value.baseName.toLowerCase());
-    }
-    if (i < patterns.length - 1) {
-      caseText += ",";
-    } else {
-      caseText += ") => ";
-    }
-  }
-  if (!(typeof expression.value === "object")) {
-    // Natively defined function
-    caseText += expression.value;
-  } else if (aliasMatches(expression.type, "MULTI_EXPRESSION")) {
-    caseText += "<multi-action>";
-  } else if (aliasMatches(expression.value.type, "CALL")) {
-    caseText += "<action>";
-  } else if (expression.value.type.baseName === "VARIABLE") {
-    caseText += expression.value.value;
-  } else {
-    caseText += charListToJsString(
-      dispatchFunction("stringify", [expression.value]),
-    );
-  }
-
-  return caseText;
-};
-
-const serializeFunction = (f) => {
-  const rankedCases = simpleRankCases(f.value);
-  return rankedCases.map((c) => serializeCase(c)).join("\n");
-};
-
-const simpleRankCases = (cases) => {
-  return cases.sort((c1, c2) => {
-    let c1Spec = 0;
-    let c2Spec = 0;
-
-    const c1Patterns = c1.value[0].value;
-    const c2Patterns = c2.value[0].value;
-
-    for (let i = 0; i < c1Patterns.length; i++) {
-      const [c1Condition, c1Type] = c1Patterns[i].value;
-      c1Spec += simpleTypeSpec(c1Type.value) * c1Condition.value[2].value;
-    }
-
-    c1Spec = c1Spec / c1Patterns.length;
-
-    for (let i = 0; i < c2Patterns.length; i++) {
-      const [c2Condition, c2Type] = c2Patterns[i].value;
-      c2Spec += simpleTypeSpec(c2Type.value) * c2Condition.value[2].value;
-    }
-
-    c2Spec = c2Spec / c2Patterns.length;
-
-    return c2Spec - c1Spec;
-  });
-};
-
-const simpleTypeSpec = (t) => {
-  if (t.baseName === "ANY") {
-    return TypeWeights.ANY;
-  }
-
-  if (t.baseName === "TUPLE" && t.types.length === 0) {
-    return TypeWeights.BASE_TUPLE;
-  }
-
-  if (isAlias(t)) {
-    return TypeWeights.NOMINAL;
-  }
-
-  return TypeWeights.EQUIVALENT;
-};
-
-export const typeToString = (t: FeverType): string => {
-  return typeToStringRec(t, "");
-};
-
-const typeToStringRec = (t, contents) => {
-  if (!t.meta) {
-    return contents + t.baseName.toLowerCase();
-  }
-
-  if (isAlias(t)) {
-    contents += t.alias;
-  }
-
-  const [open, close] = t.baseName === "LIST" ? ["[", "]"] : ["(", ")"];
-
-  const types = t.types;
-  contents += open;
-  for (let i = 0; i < types.length; i++) {
-    contents = typeToStringRec(types[i], contents);
-    if (i < types.length - 1) {
-      contents += ",";
-    }
-  }
-  contents += close;
-  return contents;
-};
-
-const namedMap = (
-  list: FeverVar,
-  action: FeverVar,
-  ctx: Context,
-  [element, index, intermediate]: string[],
-) => {
-  const internalList = [];
-  for (let i = 0; i < list.value.length; i++) {
-    const item = list.value[i];
-    const mapping = {};
-    mapping[element] = item;
-    mapping[index] = createVar(i, Primitives.NUMBER);
-    mapping[intermediate] = createVar(
-      [...internalList],
-      createTypedList(
-        internalList.length > 0 ? internalList[0].type : Primitives.ANY,
-      ),
-    );
-    const result = evaluate(
-      recreateExpressionWithVariables(action.value, mapping),
-    );
-    internalList.push(result);
-  }
-
-  const result = createVar(
-    internalList,
-    inferListType(internalList, list.type.alias),
-  );
-
-  if (isAlias(list.type)) {
-    const created = newOfType(result.type, [result], ctx);
-    if (created.type.baseName !== "ERROR") {
-      return created;
-    }
-  }
-  return result;
-};
-
-function namedMonadicMap(
-  item: FeverVar,
-  expression: FeverVar,
-  varNames: string[],
-) {
-  const variablesFromItem = dispatchFunction("lift", [item]);
-  const varMappings = variablesFromItem.value[0];
-  const mappings = {};
-  // Also this assumes there is just one concrete failure case, true for optionals but not errors
-  for (let i = 0; i < 3; i++) {
-    if (varMappings.value.length <= i) {
-      // No monad mappings registered
-      break;
-    }
-    mappings[varNames[i]] = varMappings.value[i];
-  }
-
-  let result;
-
-  if (aliasMatches(mappings[varNames[0]].type, "PASSTHROUGH")) {
-    result = mappings[varNames[0]].value;
-  } else {
-    const recreatedExpression = recreateExpressionWithVariables(
-      expression,
-      mappings,
-    );
-    result = evaluate(recreatedExpression);
-  }
-
-  return dispatchFunction("drop", [result, item]);
-}
-
-function namedMonadicFilter(
-  item: FeverVar,
-  expression: FeverVar,
-  varNames: string[],
-) {
-  const variablesFromItem = dispatchFunction("lift", [item]);
-  const varMappings = variablesFromItem.value[0];
-  const falseCaseValue = variablesFromItem.value[1];
-
-  const mappings = {};
-  // Also this assumes there is just one concrete failure case, true for optionals but not errors
-  for (let i = 0; i < 3; i++) {
-    if (varMappings.value.length <= i) {
-      // No monad mappings registered
-      break;
-    }
-    mappings[varNames[i]] = varMappings.value[i];
-  }
-
-  let result;
-
-  if (aliasMatches(mappings[varNames[0]].type, "PASSTHROUGH")) {
-    result = mappings[varNames[0]].value;
-  } else {
-    const recreatedExpression = recreateExpressionWithVariables(
-      expression,
-      mappings,
-    );
-    const boolean = evaluate(recreatedExpression);
-    if (boolean.value) {
-      // The provided item will be returned
-      result = mappings[varNames[0]];
-    } else {
-      result = falseCaseValue;
-    }
-
-    return dispatchFunction("drop", [result, item]);
-  }
-}
-
-const namedReduce = (
-  list,
-  acc,
-  expr,
-  ctx: Context,
-  [element, index, accumulator]: string[],
-  earlyTerminateIfNotFalse,
-) => {
-  const initialAcc = acc;
-  for (let i = 0; i < list.value.length; i++) {
-    const item = list.value[i];
-    const mapping = {};
-    mapping[element] = item;
-    mapping[index] = createVar(i, Primitives.NUMBER);
-    mapping[accumulator] = acc;
-    acc = evaluate(recreateExpressionWithVariables(expr, mapping));
-    // Maybe this will be slow but it seems ok
-    if (
-      earlyTerminateIfNotFalse &&
-      !dispatchFunction("==", [initialAcc, acc]).value
-    ) {
-      return acc;
-    }
-  }
-  return acc;
-};
-
-const namedFilter = (
-  list,
-  action,
-  ctx: Context,
-  [element, index, intermediate]: string[],
-  funcRef,
-) => {
-  const internalList = [];
-  for (let i = 0; i < list.value.length; i++) {
-    const item = list.value[i];
-    const mapping = {};
-    mapping[element] = item;
-    mapping[index] = createVar(i, Primitives.NUMBER);
-    mapping[intermediate] = createVar(
-      [...internalList],
-      createTypedList(
-        internalList.length > 0 ? internalList[0].type : Primitives.ANY,
-      ),
-    );
-    let result;
-    if (funcRef) {
-      // Hmmm...how do we supply vars here?
-      result = callFunctionByReference(funcRef, [item], ctx, "lambda");
-    } else {
-      result = evaluate(recreateExpressionWithVariables(action, mapping).value);
-    }
-    if (result.value) {
-      internalList.push(item);
-    }
-  }
-
-  const result = createVar(
-    internalList,
-    inferListType(internalList, list.type.alias),
-  );
-
-  if (isAlias(list.type)) {
-    const created = newOfType(result.type, [result], ctx);
-    if (created.type.baseName !== "ERROR") {
-      return created;
-    }
-  }
-  return result;
-};
-
-const isNumeric = (n) => {
-  return !isNaN(parseFloat(n)) && isFinite(n);
-};
-
-function addFunctionCase(name: FeverVar, func: FeverVar, ctx: Context) {
-  {
-    const signature = func.value[0];
-    const expression = func.value[1];
-    const size = signature.value.length;
-
-    const types = typesFromSignature(signature);
-    const conditions = conditionsFromSignature(signature);
-    const names = namesFromSignature(signature);
-    const specificities = specificitiesFromSignature(signature);
-
-    const operation = (args, ctx) => {
-      const mapping = {};
-      for (let i = 0; i < args.length; i++) {
-        mapping[names[i]] = args[i];
-      }
-      if (aliasMatches(expression.type, "MULTI_EXPRESSION")) {
-        const recreatedMultiExpression = createVar(
-          expression.value.map((line: FeverVar) =>
-            recreateExpressionWithVariables(line, mapping),
-          ),
-          Meta.MULTI_EXPRESSION,
-        );
-        return evaluate(recreatedMultiExpression);
-      } else {
-        return evaluate(
-          recreateExpressionWithVariables(expression, mapping).value,
-        );
-      }
-    };
-    return registerNewFunction(
-      name.value,
-      ctx,
-      newFunction(size, types, operation, {
-        conditions: conditions,
-        specificities: specificities,
-      }),
-      func,
-    );
-  }
-}
