+++ package.json
@@ -2,7 +2,7 @@
   "name": "fever-lang",
   "version": "1.0.0",
   "description": "",
-  "main": "fever.js",
+  "main": "Fever.ts",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
